four_bench/pyproject.toml
[tool.poetry]
name = "four-bench"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
duckdb = "0.9.1"
matplotlib = "^3.8.2"
pandas = "^2.1.3"
sqlalchemy = "^2.0.23"
tqdm = "^4.66.1"
seaborn = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
